move StoreProvider
