+++ frontend/src/utils/Providers.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { AuthProvider } from "react-oidc-context";
 import { oidcConfig } from "@/utils/oidcConfig";
 import { YMaps } from "@pbe/react-yandex-maps";
-import { StoreProvider } from "@/components/Tracking/StoreProvider";
+import { StoreProvider } from "@/components/StoreProvider/StoreProvider";
 
 export const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
